+++ lib/koolbase_flutter.dart
@@ -5,4 +5,7 @@ export 'src/auth/auth_models.dart';
 export 'src/auth/auth_exceptions.dart';
 export 'src/auth/auth_client.dart';
 export 'src/storage/storage_client.dart';
+export 'src/database/database_client.dart';
+export 'src/database/database_models.dart';
+export 'src/database/database_query.dart';
 export 'src/storage/storage_models.dart';